+++ src/Rocketeer/Rocketeer.php
@@ -26,7 +26,7 @@ class Rocketeer
 	 *
 	 * @var string
 	 */
-	const VERSION = '2.0.4';
+	const VERSION = '2.1.0';
 
 	/**
 	 * Global local mode